Fixing docs for kinds
and also, moving all the stuff about kinds: from utils to utils_kinds
